% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diskers.R
\docType{package}
\name{diskers}
\alias{diskers}
\title{recruitR}
\description{
R package to assess seeds recruitment.
}
\author{
Kevin Cazelles \href{mailto:kevin.cazelles@gmail.com}{kevin.cazelles@gmail.com}
}
