% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kern_gaussian}
\alias{kern_2Dt}
\alias{kern_exponential}
\alias{kern_exponentiel_power}
\alias{kern_gaussian}
\alias{kern_lognormal}
\title{Kernels functions}
\usage{
kern_gaussian(dist, scal)

kern_exponential(dist, scal)

kern_exponentiel_power(dist, scal, shap)

kern_2Dt(dist, scal, shap)

kern_lognormal(dist, scal, shap)
}
\arguments{
\item{distance:}{a real positive number giving the dispersal distance.}

\item{scal:}{scale parameter a real positive number giving the shape parameter of the Kernel.}

\item{shap:}{shape parameter a real positve number used as the power of the kernel.}

\item{opt:}{a real positive number related with the optimal sistance for dispersion.}
}
\description{
Common seeds isotropic dispersal kernels.
}
\section{Functions}{
\itemize{
\item \code{kern_gaussian}: Gaussian kernel.

\item \code{kern_exponential}: Exponential kernel.

\item \code{kern_exponentiel_power}: Dispersal kernels from the exponential-power family.

\item \code{kern_2Dt}: 2Dt kernel.

\item \code{kern_lognormal}: Lognormal kernel.
}}
\examples{
seqx <- seq(0,10,0.001)
plot(seqx, kern_lognormal(seqx,3,.2), type='l')
}
\author{
Kevin Cazelles
}

