% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanDispDist.R
\name{meanDispDist}
\alias{meanDispDist}
\title{Mean Dispersal Ditances}
\usage{
meanDispDist(name, scal, shap = NULL)
}
\arguments{
\item{name}{a character names of the dispersal kernel chosen.}

\item{scal}{a real positive number giving the scale parameter of the kernel.}

\item{shap}{a real positive number giving the shape parameter of the kernel.}
}
\value{
The mean dispersal distance.
}
\description{
Mean dispersal distance for five disersal kernel commonly used to model seeds dispersal.
}
\details{
Currently valid names are \code{gaussian}, \code{exponential},
\code{exponential_power}, \code{k2Dt} and \code{lognormal}.
}
\examples{
seqx <- seq(0,10,0.01)
plot(seqx, kern_lognormal(seqx,3,.2), type='l')
}
\author{
Kevin Cazelles
}
\references{
Nathan, R., Klein, E., Robledo-Arnuncio, J.J. and Revilla, E. (2012) 
Dipsersal Kernels: Review in _Dispersal Ecology and Evolution_. Oxford University Press.
}

