% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kern_gaussian}
\alias{kern_2Dt}
\alias{kern_exponential}
\alias{kern_exponential_power}
\alias{kern_gaussian}
\alias{kern_lognormal}
\title{Kernels functions}
\usage{
kern_gaussian(dist, scal)

kern_exponential(dist, scal)

kern_exponential_power(dist, scal, shap)

kern_2Dt(dist, scal, shap)

kern_lognormal(dist, scal, shap)
}
\arguments{
\item{dist}{a real positive number giving the dispersal distance.}

\item{scal}{a real positive number giving the scale parameter of the kernel.}

\item{shap}{a real positive number giving the shape parameter of the kernel.}
}
\value{
The pdf values at distance \code{dist}.
}
\description{
Common seed isotropic dispersal kernels.
}
\details{
This is a set of five disersal kernel commonly used to model seeds dispersal.
}
\section{Functions}{
\itemize{
\item \code{kern_gaussian}: Gaussian kernel.
\deqn{k(dist, scal) = \frac{1}{pi*scal^2}
exp\left(-\left(\frac{dist}{scal}\right)^2\right)}

\item \code{kern_exponential}: Exponential kernel.

\item \code{kern_exponential_power}: Dispersal kernels from the exponential-power family.

\item \code{kern_2Dt}: 2Dt kernel.

\item \code{kern_lognormal}: Lognormal kernel.
}}
\examples{
seqx <- seq(0,10,0.01)
plot(seqx, kern_lognormal(seqx,3,.2), type='l')
}
\author{
Kevin Cazelles
}

