% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernels}
\alias{kern_2Dt}
\alias{kern_exponential}
\alias{kern_exponential_power}
\alias{kern_gaussian}
\alias{kern_lognormal}
\alias{kernels}
\title{Common seed isotropic dispersal kernels.}
\usage{
kernels()

kern_gaussian(dist, scal)

kern_exponential(dist, scal)

kern_exponential_power(dist, scal, shap)

kern_2Dt(dist, scal, shap)

kern_lognormal(dist, scal, shap)
}
\arguments{
\item{dist}{a real positive number giving the dispersal distance.}

\item{scal}{a real positive number giving the scale parameter of the kernel.}

\item{shap}{a real positive number giving the shape parameter of the kernel.}
}
\value{
The pdf values at distance \code{dist}.
}
\description{
This is a set of five disersal kernel commonly used to model seeds dispersal.
}
\section{Functions}{
\itemize{
\item \code{kern_gaussian}: Exponential kernel.

\item \code{kern_exponential}: Exponential kernel.

\item \code{kern_exponential_power}: Dispersal kernels from the exponential-power family.

\item \code{kern_2Dt}: 2Dt kernel.

\item \code{kern_lognormal}: Lognormal kernel.
}}
\examples{
seqx <- seq(0,10,0.01)
plot(seqx, kern_lognormal(seqx,3,.2), type='l')
}
\author{
Kevin Cazelles
}
\references{
Nathan, R., Klein, E., Robledo-Arnuncio, J. J. and Revilla, E. (2012).
\code{Dispersal kernels: review} in Dispersal Ecology and Evolution,
186–210. Oxford University Press.
}

