% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/meanDispDist.R
\name{kernels}
\alias{kernels}
\alias{kern_gaussian}
\alias{kern_exponential}
\alias{kern_exponential_power}
\alias{kern_2Dt}
\alias{kern_lognormal}
\alias{meanDispDist_gaussian}
\alias{meanDispDist_exponential}
\alias{meanDispDist_exponential_power}
\alias{meanDispDist_2Dt}
\alias{meanDispDist_lognormal}
\title{Common seed isotropic dispersal kernels.}
\usage{
kernels()

kern_gaussian(dist, scal)

kern_exponential(dist, scal)

kern_exponential_power(dist, scal, shap)

kern_2Dt(dist, scal, shap)

kern_lognormal(dist, scal, shap)

meanDispDist_gaussian(dist, scal)

meanDispDist_exponential(dist, scal)

meanDispDist_exponential_power(dist, scal, shap)

meanDispDist_2Dt(dist, scal, shap)

meanDispDist_lognormal(dist, scal, shap)
}
\arguments{
\item{dist}{a real positive number giving the dispersal distance.}

\item{scal}{a real positive number giving the scale parameter of the kernel.}

\item{shap}{a real positive number giving the shape parameter of the kernel.}
}
\value{
The pdf values at distance \code{dist}.
}
\description{
This is a set of five disersal kernel commonly used to model seeds dispersal.
}
\section{Functions}{
\itemize{
\item \code{kern_gaussian}: Exponential kernel.

\item \code{kern_exponential}: Exponential kernel.

\item \code{kern_exponential_power}: Dispersal kernels from the exponential-power family.

\item \code{kern_2Dt}: 2Dt kernel.

\item \code{kern_lognormal}: Lognormal kernel.

\item \code{meanDispDist_gaussian}: Exponential kernel.

\item \code{meanDispDist_exponential}: Exponential kernel.

\item \code{meanDispDist_exponential_power}: Dispersal kernels from the exponential-power family.

\item \code{meanDispDist_2Dt}: 2Dt kernel.

\item \code{meanDispDist_lognormal}: Lognormal kernel.
}}

\examples{
seqx <- seq(0,10,0.01)
plot(seqx, kern_lognormal(seqx,3,.2), type='l')
}
\references{
Nathan, R., Klein, E., Robledo-Arnuncio, J.J. and Revilla, E. (2012)
Dipsersal Kernels: Review in \emph{Dispersal Ecology and Evolution}. Oxford University Press.
}
\author{
Kevin Cazelles
}
