% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanDispDist.R
\name{meanDispDist}
\alias{meanDispDist}
\title{Mean dispersal distance}
\usage{
meanDispDist(name, scal, shap = NULL)
}
\arguments{
\item{name}{a character string that names of the dispersal kernel chosen.}

\item{scal}{a real positive number giving the scale parameter of the kernel.}

\item{shap}{a real positive number giving the shape parameter of the kernel.}
}
\value{
The mean dispersal distance.
}
\description{
Mean dispersal distance for the five disersal kernel implemented.
}
\details{
Currently, valid names are \code{gaussian}, \code{exponential},
\code{exponential_power}, \code{k2Dt} and \code{lognormal}.
}
\examples{
meanDispDist('gaussian', 10)

}
\references{
Nathan, R., Klein, E., Robledo-Arnuncio, J.J. and Revilla, E. (2012)
Dipsersal Kernels: Review in \emph{Dispersal Ecology and Evolution}. Oxford University Press.
}
\author{
Kevin Cazelles
}
